+++ src/messages/error.rs
@@ -1,0 +1,40 @@
+/// An error returned when trying to use [TryFrom] on a [Message] to a specific [message][msg]
+///
+/// [TryFrom]: https://doc.rust-lang.org/std/convert/trait.TryFrom.html
+/// [Message]: ../decode/struct.Message.html
+/// [msg]: ./messages/index.html
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum InvalidMessage {
+    /// An invalid command was found for this message
+    InvalidCommand {
+        /// Expected this command
+        expected: String,
+        /// Got this command
+        got: String,
+    },
+    /// Expected a nickname attached to this message
+    ExpectedNick,
+    /// Expected an argument at a position in this message
+    ExpectedArg {
+        /// Argument position
+        pos: usize,
+    },
+    /// Expected this message to have data attached
+    ExpectedData,
+}
+
+impl std::fmt::Display for InvalidMessage {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            InvalidMessage::InvalidCommand { expected, got } => {
+                write!(f, "invalid command. got: {} expected {}", got, expected)
+            }
+            InvalidMessage::ExpectedNick => f.write_str("expected nickname"),
+            InvalidMessage::ExpectedArg { pos } => write!(f, "expected arg at {}", pos),
+            InvalidMessage::ExpectedData => f.write_str("expected data"),
+        }
+    }
+}
+
+impl std::error::Error for InvalidMessage {}
